Add files via upload
This script reads in whole-brain raw topological metric data associated with experimental fMRI conditions. 
It outputs tables with 10,000 permutations per subject of randomized data, for use in subsequent permutation tests to evaluate statistical significance.
The tables are structured so that data can be fit by generalized linear models in statsmodels.